Moving tokenizer to DbotPreprocessTextData
