+++ Packs/ApiModules/ReleaseNotes/1_1_5.md
@@ -1,4 +1,6 @@
 
 #### Scripts
-##### New: MLApiModule
-A piece of ML code that is used and which is shared by different ML automations.
+##### CSVFeedApiModule
+- Maintenance and stability enhancements.
+- Added the `limit`, `value_field` parameters to the fetch indicators command.
+- Upgraded the docker image to jmespath:1.0.0.12410
