--- Packs/ApiModules/Scripts/MLApiModule/MLApiModule.yml
@@ -1,14 +1,0 @@
-commonfields:
-  id: MLApiModule
-  version: -1
-name: MLApiModule
-script: '-'
-subtype: python3
-system: true
-timeout: '0'
-type: python
-dockerimage: demisto/ml:1.0.0.12420
-comment: A piece of ML code that is used and which is shared by different ML automations.
-tags:
-- ml
-fromversion: "5.0.0"
